Add script for warning and hibernating.
This is a simple script I wrote that will hopefully work on most distros
for:

A) Warning the user that the battery level is critical.
B) Automating the power state change process (optional, on by default).

Tested on Slackware 13.37 x86_64 with 3.2.7 vanilla kernel.
Tested with Xdialog 2.3.1 and xmessage 1.0.3.

To use with batterymon, start batterymon with the `--on-critical'
option pointed at this script.